update tests, move to it's own file
